+++ src/components/App/styles.scss
@@ -6,7 +6,7 @@ body {
 }
 
 .layout {
-  height: 100vh;
+  min-height: 100vh;
   display: flex;
   flex-direction: column;
 }
